Remove title tooltips (#115)
This change replaces the `title` attribute with `label` for the ability-pick and ability-tree custom elements in order to remove the default tooltips created by the browser. The `label` attribute of the ability-tree custom element is not currently used by its code but it is kept just in case it might be needed in the future.

Closes #114