Add Arcanistics ability tree (preview based on devlog) (#122)
Add the Arcanistics ability tree based on the information provided in [Devlog: Arcanistics](https://store.steampowered.com/news/app/625960/view/514085449373844395).

This is an early preview as the final tree may change upon release.

Once the Arcanistics tree is officially added to the game, tooltip descriptions and ability parameters will be updated to match the in-game values.

Closes #121